TEIIDTOOLS-723 consolidating server packages
