+++ server/komodo-server/src/main/java/org/komodo/metadata/internal/TeiidServer.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.komodo.rest;
+package org.komodo.metadata.internal;
 
 import java.util.concurrent.ConcurrentHashMap;
