+++ server/komodo-server/src/main/java/org/komodo/rest/relational/PublishRequestPayload.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.komodo.rest.relational.request;
+package org.komodo.rest.relational;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonInclude.Include;
