+++ server/komodo-server/src/main/java/org/komodo/rest/relational/RestDataservice.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.komodo.rest.relational.dataservice;
+package org.komodo.rest.relational;
 
 import java.util.Arrays;
