+++ server/komodo-server/src/main/java/org/komodo/rest/relational/RestSyndesisSourceStatus.java
@@ -1,4 +1,4 @@
-package org.komodo.rest.relational.response.metadata;
+package org.komodo.rest.relational;
 
 import java.util.Arrays;
 import java.util.List;
