+++ server/komodo-server/src/main/java/org/komodo/rest/service/KomodoDataserviceService.java
@@ -54,9 +54,9 @@ import org.komodo.openshift.TeiidOpenShiftClient;
 import org.komodo.rest.KomodoRestV1Application.V1Constants;
 import org.komodo.rest.KomodoService;
 import org.komodo.rest.ResourceNotFound;
+import org.komodo.rest.relational.KomodoStatusObject;
 import org.komodo.rest.relational.RelationalMessages;
-import org.komodo.rest.relational.dataservice.RestDataservice;
-import org.komodo.rest.relational.response.KomodoStatusObject;
+import org.komodo.rest.relational.RestDataservice;
 import org.komodo.utils.StringNameValidator;
 import org.komodo.utils.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
