+++ server/komodo-server/src/main/java/org/komodo/rest/service/KomodoMetadataService.java
@@ -64,14 +64,14 @@ import org.komodo.openshift.TeiidOpenShiftClient;
 import org.komodo.rest.AuthHandlingFilter.OAuthCredentials;
 import org.komodo.rest.KomodoRestV1Application.V1Constants;
 import org.komodo.rest.KomodoService;
+import org.komodo.rest.relational.KomodoQueryAttribute;
+import org.komodo.rest.relational.KomodoStatusObject;
+import org.komodo.rest.relational.PublishRequestPayload;
 import org.komodo.rest.relational.RelationalMessages;
+import org.komodo.rest.relational.RestSyndesisSourceStatus;
+import org.komodo.rest.relational.RestViewSourceInfo;
 import org.komodo.rest.relational.connection.RestSchemaNode;
 import org.komodo.rest.relational.connection.RestSourceSchema;
-import org.komodo.rest.relational.request.KomodoQueryAttribute;
-import org.komodo.rest.relational.request.PublishRequestPayload;
-import org.komodo.rest.relational.response.KomodoStatusObject;
-import org.komodo.rest.relational.response.metadata.RestSyndesisSourceStatus;
-import org.komodo.rest.relational.response.vieweditorstate.RestViewSourceInfo;
 import org.komodo.utils.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
