+++ server/komodo-server/src/main/java/org/komodo/rest/service/KomodoUtilService.java
@@ -44,9 +44,9 @@ import org.komodo.metadata.MetadataInstance.ValidationResult;
 import org.komodo.rest.KomodoRestV1Application;
 import org.komodo.rest.KomodoRestV1Application.V1Constants;
 import org.komodo.rest.KomodoService;
+import org.komodo.rest.relational.KomodoStatusObject;
 import org.komodo.rest.relational.RelationalMessages;
-import org.komodo.rest.relational.response.KomodoStatusObject;
-import org.komodo.rest.relational.response.vieweditorstate.RestViewDefinitionStatus;
+import org.komodo.rest.relational.RestViewDefinitionStatus;
 import org.komodo.utils.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
