--- server/komodo-server/src/test/java/org/komodo/rest/relational/json/AbstractSerializerTest.java
@@ -1,30 +1,0 @@
-/*
- * Copyright Red Hat, Inc. and/or its affiliates
- * and other contributors as indicated by the @author tags and
- * the COPYRIGHT.txt file distributed with this work.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.komodo.rest.relational.json;
-
-import org.komodo.rest.json.JsonConstants;
-
-/**
- *
- */
-@SuppressWarnings( {"javadoc", "nls" })
-public abstract class AbstractSerializerTest implements JsonConstants {
-
-    protected static final String DATASERVICE_NAME = "dataservice1";
-
-}
